+++ apps/backoffice-v2/src/common/utils/is-pdf/is-pdf.ts
@@ -1,0 +1,4 @@
+export const isPdf = (file: { imageUrl: string; fileType: string }) =>
+  file?.fileType === 'application/pdf' ||
+  file?.fileType === 'pdf' ||
+  file?.imageUrl?.endsWith('.pdf');