+++ services/workflows-service/src/storage/get-file-storage-manager.ts
@@ -68,19 +68,19 @@ export const downloadFileFromS3 = async (
 export const createPresignedUrlWithClient = async ({
   bucketName,
   fileNameInBucket,
-  fileTypeByEnding,
+  mimeType,
   service = 'cloudfront',
 }: {
   bucketName: string;
   fileNameInBucket: string;
-  fileTypeByEnding?: string;
+  mimeType?: string;
   service?: 's3' | 'cloudfront';
 }): Promise<TLocalFile> => {
   if (
     service === 'cloudfront' &&
     process.env.AWS_S3_CF_URL &&
     process.env.AWS_S3_CF_KEYPAIR_ID &&
-    (process.env.AWS_S3_CF_PRIVATE_KEY || process.env.AWS_S3_CF_PRIVATE_KEY_BASE64)
+    (process.env.AWS_S3_CF_PRIVATE_KEY_BASE64 || process.env.AWS_S3_CF_PRIVATE_KEY)
   ) {
     return cloudfrontPresignedUrl({
       fileNameInBucket,
@@ -90,24 +90,24 @@ export const createPresignedUrlWithClient = async ({
   return s3PresignedUrl({
     bucketName,
     fileNameInBucket,
-    fileTypeByEnding,
+    mimeType,
   });
 };
 
 export const s3PresignedUrl = async ({
   bucketName,
   fileNameInBucket,
-  fileTypeByEnding,
+  mimeType,
 }: {
   bucketName: string;
   fileNameInBucket: string;
-  fileTypeByEnding?: string;
+  mimeType?: string;
 }) => {
   const s3Client = new S3Client(AwsS3FileConfig.fetchClientConfig(process.env));
   const command = new GetObjectCommand({
     Bucket: bucketName,
     Key: fileNameInBucket,
-    ResponseContentType: fileTypeByEnding && `application/${fileTypeByEnding}`,
+    ResponseContentType: mimeType ? mimeType : undefined,
   });
 
   return getSignedUrl(s3Client, command, { expiresIn: 1800 });