+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ballerine/workflow-core
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.2
+
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.1
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.0
+
 ## 0.5.0
 
 ### Minor Changes