+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.2
+  - @ballerine/workflow-core@0.5.3
+
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.1
+  - @ballerine/workflow-core@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.0
+  - @ballerine/workflow-core@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes